backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
supabase==2.23.0
PyJWT==2.8.0
cryptography==41.0.7
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0
google-api-python-client==2.108.0
Werkzeug==3.0.1
marshmallow==3.20.1
websockets>=15.0
requests==2.31.0
pyjwt[crypto]>=2.8.0
